Convert deleteEmptyThought action-creator to reducer.
