+++ src/action-creators/index.js
@@ -4,7 +4,6 @@ export { default as cursorNext } from './cursorNext'
 export { default as cursorPrev } from './cursorPrev'
 export { default as dataIntegrityCheck } from './dataIntegrityCheck'
 export { default as deleteAttribute } from './deleteAttribute'
-export { default as deleteEmptyThought } from './deleteEmptyThought'
 export { default as error } from './error'
 export { default as expandContextThought } from './expandContextThought'
 export { default as flushQueue } from './flushQueue'
